requirements.txt
langchain==0.2.10
langchain-openai==0.1.17
python-dotenv==1.0.1
langchain-groq==0.1.6
langchain-community==0.2.10
bs4==0.0.2
unstructured==0.15.0
pypdf==4.3.1
wikipedia==1.4.0
langchain-chroma==0.1.2
langserve==0.2.2
langgraph==0.1.19
flask==3.0.2
langchain_core
httpx==0.24.1
streamlit